Employee List WORKING AND EDITING
The employee list has a problem, it wont update the project. It stays the same.

My Data does the same as the employeelist edit, it wont save, got figure that out.

Need to make the login page sessions attributes to mydata
Need to Make the employee data/admin project change.
Need to get the sessions attributes for My Data.